fix: rename folder name to 'Metanodes'
