+++ src/modules/scenes/Main/Metanodes/BondToLiquidity/styled.tsx
