+++ src/modules/scenes/Main/Metanodes/BondToLiquidity2/index.tsx
