+++ src/modules/scenes/Main/Metanodes/BondToLiquidity2/styled.tsx
