+++ src/modules/scenes/Main/Metanodes/BridgeMetanodes/styled.tsx
