+++ src/modules/scenes/Main/Metanodes/BridgeMobileMetanodes/styled.tsx
