+++ src/modules/scenes/Main/Metanodes/Earnings/index.tsx
