+++ src/modules/scenes/Main/Metanodes/Earnings/styled.tsx
