+++ src/modules/scenes/Main/Metanodes/GeneralInfo/index.tsx
