+++ src/modules/scenes/Main/Metanodes/GeneralInfo/styled.tsx
