+++ src/modules/scenes/Main/Metanodes/LiquidityRatio/styled.tsx
