+++ src/modules/scenes/Main/Metanodes/MeatanodeInfo/index.tsx
