+++ src/modules/scenes/Main/Metanodes/MeatanodeInfo/styled.tsx
