+++ src/modules/scenes/Main/Metanodes/NodeStatus/styled.tsx
