+++ src/modules/scenes/Main/Metanodes/TotalNodes/styled.tsx
