+++ src/modules/scenes/Main/index.tsx
@@ -5,7 +5,7 @@ import { ExplorerMain } from '../Common';
 export { Browser } from './Explorer';
 export { BrowserDetail } from './Swap';
 export { BrowserPool } from './Pool';
-export { BrowserMetanodes } from './NewMetanodes';
+export { BrowserMetanodes } from './Metanodes';
 export { BrowserFees } from './Fees/';
 export { BrowserAsset } from './Asset';
 export { MetanodeEarners } from './MetanodeEarners';
